refactor(mla): move mla_fp8 module into models/deepseek_vl_v2
Review nit from #1284: the FP8-MLA helpers are only used by the
deepseek_vl_v2 model, so house them under its package instead of the
mlx_vlm root. Pure move + import-path updates; no behavior change
(51/51 tests pass).
